requirements.txt
rich~=13.7.0
ultralytics~=8.1.9
pillow~=10.2.0